+++ Oracle.NoSQL.SDK/src/NsonProtocol/RequestSerializer.Query.cs
@@ -360,6 +360,8 @@ namespace Oracle.NoSQL.SDK.NsonProtocol
                 WriteDurability(writer, request.Durability);
             }
 
+            WriteLastWriteMetadata(writer, request.LastWriteMetadata);
+
             OptionallyWriteInt32(writer, FieldNames.MaxReadKB,
                 request.Options?.MaxReadKB);
             OptionallyWriteInt32(writer, FieldNames.MaxWriteKB,